DISPATCH-606 - Default maxSessionFrames to 0.  Zero now means the maximum possible session window (2^31-1 on 64-buit arches and (2^31-1)/max_frame_size on 32-bit architectures.  The latter restriction is a result of a limitation in the Proton API.

+++ include/qpid/dispatch/server.h
@@ -444,7 +444,7 @@ typedef struct qd_server_config_t {
      * to equal the specified sessionMaxFrames value measured in units of transfer frames.
      * This calculation is done to satisfy proton pn_session_set_incoming_capacity().
      */
-    uint32_t incoming_capacity;
+    size_t incoming_capacity;
 
     /**
      * The idle timeout, in seconds.  If the peer sends no data frames in this many seconds, the
